requirements.txt
# ── Core ────────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Claude API ───────────────────────────────────────────────────────────────
anthropic==0.28.0

# ── ElevenLabs TTS ───────────────────────────────────────────────────────────
requests==2.32.3
elevenlabs==1.5.0

# ── Whisper (sous-titres) ────────────────────────────────────────────────────
openai-whisper==20231117
torch==2.3.1
torchaudio==2.3.1
numba==0.60.0

# ── Playwright (upload automatisé) ───────────────────────────────────────────
playwright==1.44.0

# ── Utilitaires ──────────────────────────────────────────────────────────────
# Aucune dépendance supplémentaire requise — stdlib + packages listés ci-dessus
